move todo components into pagex
